+++ ChangeLog.md
@@ -5,6 +5,14 @@
 
 # CaPyCli - Clearing Automation Python Command Line Tool for SW360
 
+## NEXT
+
+* Warn if Keycloak client_id and client_secret are passed on the command line.
+  See Readme.md for recommendations how to store them.
+* Config file can be located in home directory or current working directory.
+* Document config file usage and add missing aliases so all command line
+  options can be used in the config file.
+
 ## 2.12.0
 
 * Because of security reasons `-client_id` and `-client_secret` should only